Linux Permissions fix
Removes the manual permission setting for linux, which should default it
to using the permissions of the user running the process. This will
prevent users needing unessessary access to create files.
Also adds error handling to the base worker to log errors across all
workers which should help diagnose issues in the future.